+++ src/app/services/queuemanager.service.js
@@ -17,6 +17,7 @@ angular.module("mopify.services.queuemanager", [
     // Setup websoclet
     var protocol = (typeof document !== "undefined" && document.location.protocol === "https:") ? "wss://" : "ws://";
     var ws;
+    var recoveringDelay = 1000;
 
     /**
      * Send a request to the server's queuemanager class
@@ -111,6 +112,7 @@ angular.module("mopify.services.queuemanager", [
 
         $timeout(function(){
             if(ws.readyState === 1){
+                console.info("Websocket: connection ready");
                 wsconnection = true;
                 handleWaitlist();
             }
@@ -118,6 +120,35 @@ angular.module("mopify.services.queuemanager", [
                 self.checkConnectionReady();
             }
         }, 200);
+    };
+
+    /**
+     * Start revovering the websocket
+     * @return {[type]} [description]
+     */
+    QueueManager.prototype.startRecovering = function(){
+        var that = this;
+
+        if(ws !== undefined){
+            this.closeWebsocketConnection();
+        }
+
+        $timeout(function(){
+            that.setupWebsocket();
+        }, recoveringDelay);
+
+        recoveringDelay = recoveringDelay + 1000;
+    };
+
+    /**
+     * Formarly close the current websocket connection
+     * @return {[type]} [description]
+     */
+    QueueManager.prototype.closeWebsocketConnection = function(){
+        ws.onopen = function(){};
+        ws.onclose = function(){};
+        ws.onerror = function(){};
+        ws.onmessage = function(){};
     };
 
     /**
@@ -138,18 +169,8 @@ angular.module("mopify.services.queuemanager", [
         // Set connection to false on close
         ws.onclose = function(){
             wsconnection = false;
-
-            $timeout(function(){
-                that.setupWebsocket();
-            }, 2000);
-        };
-
-        ws.onerror = function(evt) {
-            wsconnection = false;
-
-            $timeout(function(){
-                that.setupWebsocket();
-            }, 2000);
+            ws.close();
+            that.startRecovering();
         };
 
         // Handle incoming messages